pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-intel-bot"
version = "0.1.0"
description = "AI-powered crypto market intelligence Discord bot"
requires-python = ">=3.11"
authors = [{ name = "George Castillo" }]
license = { text = "MIT" }
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "discord.py>=2.3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "feedparser>=6.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "groq>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "httpx",
]

[project.scripts]
crypto-intel = "crypto_intel.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/crypto_intel"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["crypto_intel"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
